Remove modules, redo the queues and inter-thread communication
The scanner no longer loads modules; it is now just an SSL scanner.
The Wordpress module was erased, and the former SSL module extracted
into a SSLTester class, to be used as a single thread.

The queues were refactored as two dedicated classes, and scan-state
dumping/loading simplified. The loader/saver threads are gone, as well
as most global variables. Dumping state is now instant, since there is
no wait for the scans in progress to finish.

The scanner is now structured thus:

* the SSLTester module does the scanning, and provides results able to
   insert themselves into the database

* the main script spins up as many SSLTester threads as it needs, and
   handles everything else (preserving state, DB init, feeding targets
   and gathering results, etc.)